+++ backend/src/test/java/za/ac/cput/tekkiestorecapstone/controller/ShoeVariantControllerTest.java
@@ -21,6 +21,7 @@ import za.ac.cput.tekkiestorecapstone.factory.ShoeSizeFactory;
 import za.ac.cput.tekkiestorecapstone.factory.ShoeVariantFactory;
 import za.ac.cput.tekkiestorecapstone.service.ShoeVariantService;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
@@ -52,7 +53,7 @@ class ShoeVariantControllerTest {
                 "Sneaker",
                 "Iconic lifestyle sneaker",
                 "Men",
-                2499.00,
+                BigDecimal.valueOf(2499.00),
                 List.of("https://res.cloudinary.com/test/nike.jpg")
         );
